move getFileUri to FilePathTrait
